requirements.txt
Django===3.1.2
django-filter
djangorestframework
gunicorn
django-cors-headers
django-tinymce==3.4.0
djangorestframework==3.13.1
gunicorn==20.1.0
Markdown==3.4.1
Pillow==9.2.0
psycopg2-binary